requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
behave==1.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
deepdiff==5.7.0
dictor==0.1.9
distlib==0.3.4
filelock==3.6.0
idna==3.3
ordered-set==4.0.2
parse==1.19.0
parse-type==0.6.0
platformdirs==2.5.1
pluggy==1.0.0
pydantic==1.9.0
PyHamcrest==2.0.3
pyotp==2.6.0
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
